Add quarterly cup workflow, added new main.py using forecasting-tools, and updated readme
