+++ main.py
@@ -1,305 +1,98 @@
+import argparse
 import asyncio
-import datetime
-import json
+import logging
 import os
-import re
-import dotenv
-dotenv.load_dotenv()
+from datetime import datetime
+from typing import Literal
 
-from openai import AsyncOpenAI
-import numpy as np
-import requests
-import forecasting_tools
-from asknews_sdk import AskNewsSDK
+from forecasting_tools import (AskNewsSearcher, BinaryQuestion, ForecastBot,
+                               GeneralLlm, MetaculusApi, MetaculusQuestion,
+                               MultipleChoiceQuestion, NumericDistribution,
+                               NumericQuestion, PredictedOptionList,
+                               PredictionExtractor, ReasonedPrediction,
+                               SmartSearcher, clean_indents)
 
-
-######################### CONSTANTS #########################
-# Constants
-SUBMIT_PREDICTION = True  # set to True to publish your predictions to Metaculus
-USE_EXAMPLE_QUESTIONS = False  # set to True to forecast example questions rather than the tournament questions
-NUM_RUNS_PER_QUESTION = 5  # The median forecast is taken between NUM_RUNS_PER_QUESTION runs
-SKIP_PREVIOUSLY_FORECASTED_QUESTIONS = True
-GET_NEWS = False  # set to True to enable the bot to do online research
-
-# Environment variables
-# You only need *either* Exa or Perplexity or AskNews keys for online research
-METACULUS_TOKEN = os.getenv("METACULUS_TOKEN")
-PERPLEXITY_API_KEY = os.getenv("PERPLEXITY_API_KEY")
-ASKNEWS_CLIENT_ID = os.getenv("ASKNEWS_CLIENT_ID")
-ASKNEWS_SECRET = os.getenv("ASKNEWS_SECRET")
-EXA_API_KEY = os.getenv("EXA_API_KEY")
-OPENAI_API_KEY = os.getenv("OPENAI_API_KEY") # You'll also need the OpenAI API Key if you want to use the Exa Smart Searcher
-
-# The tournament IDs below can be used for testing your bot.
-Q4_2024_AI_BENCHMARKING_ID = 32506
-Q1_2025_AI_BENCHMARKING_ID = 32627
-Q4_2024_QUARTERLY_CUP_ID = 3672
-Q1_2025_QUARTERLY_CUP_ID = 32630
-AXC_2025_TOURNAMENT_ID = 32564
-GIVEWELL_ID = 3600
-RESPIRATORY_OUTLOOK_ID = 3411
-
-TOURNAMENT_ID = Q1_2025_AI_BENCHMARKING_ID
-
-# The example questions can be used for testing your bot. (note that question and post id are not always the same)
-EXAMPLE_QUESTIONS = [  # (question_id, post_id)
-    (578, 578),  # Human Extinction - Binary - https://www.metaculus.com/questions/578/human-extinction-by-2100/
-    (14333, 14333),  # Age of Oldest Human - Numeric - https://www.metaculus.com/questions/14333/age-of-oldest-human-as-of-2100/
-    (22427, 22427),  # Number of New Leading AI Labs - Multiple Choice - https://www.metaculus.com/questions/22427/number-of-new-leading-ai-labs/
-]
-
-# Also, we realize the below code could probably be cleaned up a bit in a few places
-# Though we are assuming most people will dissect it enough to make this not matter much
-# Hopefully this is a good starting point for people to build on and get a gist of whats involved
-
-######################### HELPER FUNCTIONS #########################
-
-# @title Helper functions
-AUTH_HEADERS = {"headers": {"Authorization": f"Token {METACULUS_TOKEN}"}}
-API_BASE_URL = "https://www.metaculus.com/api"
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
 
 
-def post_question_comment(post_id: int, comment_text: str) -> None:
+class Q1TemplateBot(ForecastBot):
     """
-    Post a comment on the question page as the bot user.
+    This is a template bot that uses the forecasting-tools library to simplify bot making.
+
+    The main entry point of this bot is `forecast_on_tournament` in the parent class.
+    However generally the flow is:
+    - Load questions from Metaculus
+    - For each question
+        - Execute run_research for research_reports_per_question runs
+        - Execute respective run_forecast function for `predictions_per_research_report * research_reports_per_question` runs
+        - Aggregate the predictions
+        - Submit prediction (if publish_reports_to_metaculus is True)
+    - Return a list of ForecastReport objects
+
+    Only the research and forecast functions need to be implemented in ForecastBot subclasses.
+
+    If you end up having trouble with rate limits and want to try a more sophisticated rate limiter try:
+    ```
+    from forecasting_tools.ai_models.resource_managers.refreshing_bucket_rate_limiter import RefreshingBucketRateLimiter
+    rate_limiter = RefreshingBucketRateLimiter(
+        capacity=2,
+        refresh_rate=1,
+    ) # Allows 1 request per second on average with a burst of 2 requests initially. Set this as a class variable
+    await self.rate_limiter.wait_till_able_to_acquire_resources(1) # 1 because it's consuming 1 request (use more if you are adding a token limit)
+    ```
+
+    Check out https://github.com/Metaculus/forecasting-tools for a full set of features from this package.
+    Most notably there is a built in benchmarker that integrates with ForecastBot objects.
     """
 
-    response = requests.post(
-        f"{API_BASE_URL}/comments/create/",
-        json={
-            "text": comment_text,
-            "parent": None,
-            "included_forecast": True,
-            "is_private": True,
-            "on_post": post_id,
-        },
-        **AUTH_HEADERS,  # type: ignore
-    )
-    if not response.ok:
-        raise RuntimeError(response.text)
+    _max_concurrent_questions = 2  # Set this to whatever works for your search-provider/ai-model rate limits
+    _concurrency_limiter = asyncio.Semaphore(_max_concurrent_questions)
 
-
-def post_question_prediction(question_id: int, forecast_payload: dict) -> None:
-    """
-    Post a forecast on a question.
-    """
-    url = f"{API_BASE_URL}/questions/forecast/"
-    response = requests.post(
-        url,
-        json=[
-            {
-                "question": question_id,
-                **forecast_payload,
-            },
-        ],
-        **AUTH_HEADERS,  # type: ignore
-    )
-    print(f"Prediction Post status code: {response.status_code}")
-    if not response.ok:
-        raise RuntimeError(response.text)
-
-
-def create_forecast_payload(
-    forecast: float | dict[str, float] | list[float],
-    question_type: str,
-) -> dict:
-    """
-    Accepts a forecast and generates the api payload in the correct format.
-
-    If the question is binary, forecast must be a float.
-    If the question is multiple choice, forecast must be a dictionary that
-      maps question.options labels to floats.
-    If the question is numeric, forecast must be a dictionary that maps
-      quartiles or percentiles to datetimes, or a 201 value cdf.
-    """
-    if question_type == "binary":
-        return {
-            "probability_yes": forecast,
-            "probability_yes_per_category": None,
-            "continuous_cdf": None,
-        }
-    if question_type == "multiple_choice":
-        return {
-            "probability_yes": None,
-            "probability_yes_per_category": forecast,
-            "continuous_cdf": None,
-        }
-    # numeric or date
-    return {
-        "probability_yes": None,
-        "probability_yes_per_category": None,
-        "continuous_cdf": forecast,
-    }
-
-
-def list_posts_from_tournament(
-    tournament_id: int = TOURNAMENT_ID, offset: int = 0, count: int = 50
-) -> list[dict]:
-    """
-    List (all details) {count} posts from the {tournament_id}
-    """
-    url_qparams = {
-        "limit": count,
-        "offset": offset,
-        "order_by": "-hotness",
-        "forecast_type": ",".join(
-            [
-                "binary",
-                "multiple_choice",
-                "numeric",
-            ]
-        ),
-        "tournaments": [tournament_id],
-        "statuses": "open",
-        "include_description": "true",
-    }
-    url = f"{API_BASE_URL}/posts/"
-    response = requests.get(url, **AUTH_HEADERS, params=url_qparams)  # type: ignore
-    if not response.ok:
-        raise Exception(response.text)
-    data = json.loads(response.content)
-    return data
-
-
-def get_open_question_ids_from_tournament() -> list[tuple[int, int]]:
-    posts = list_posts_from_tournament()
-
-    post_dict = dict()
-    for post in posts["results"]:
-        if question := post.get("question"):
-            # single question post
-            post_dict[post["id"]] = [question]
-
-    open_question_id_post_id = []  # [(question_id, post_id)]
-    for post_id, questions in post_dict.items():
-        for question in questions:
-            if question.get("status") == "open":
-                print(
-                    f"ID: {question['id']}\nQ: {question['title']}\nCloses: "
-                    f"{question['scheduled_close_time']}"
+    async def run_research(self, question: MetaculusQuestion) -> str:
+        async with self._concurrency_limiter:
+            research = ""
+            if os.getenv("ASKNEWS_CLIENT_ID") and os.getenv("ASKNEWS_SECRET"):
+                research = AskNewsSearcher().get_formatted_news(
+                    question.question_text
                 )
-                open_question_id_post_id.append((question["id"], post_id))
+            elif os.getenv("EXA_API_KEY"):
+                research = await self._call_exa_smart_searcher(
+                    question.question_text
+                )
+            elif os.getenv("PERPLEXITY_API_KEY"):
+                research = await self._call_perplexity(question.question_text)
+            else:
+                research = ""
+            logger.info(f"Found Research for {question.page_url}:\n{research}")
+            return research
 
-    return open_question_id_post_id
-
-
-def get_post_details(post_id: int) -> dict:
-    """
-    Get all details about a post from the Metaculus API.
-    """
-    url = f"{API_BASE_URL}/posts/{post_id}/"
-    print(f"Getting details for {url}")
-    response = requests.get(
-        url,
-        **AUTH_HEADERS,  # type: ignore
-    )
-    if not response.ok:
-        raise Exception(response.text)
-    details = json.loads(response.content)
-    return details
-
-CONCURRENT_REQUESTS_LIMIT = 5
-llm_rate_limiter = asyncio.Semaphore(CONCURRENT_REQUESTS_LIMIT)
-
-
-async def call_llm(prompt: str, model: str = "gpt-4o", temperature: float = 0.3) -> str:
-    """
-    Makes a streaming completion request to OpenAI's API with concurrent request limiting.
-    """
-
-    # Remove the base_url parameter to call the OpenAI API directly
-    # Also checkout the package 'litellm' for one function that can call any model from any provider
-    # Email support@metaculus.com if you need credit for the Metaculus OpenAI/Anthropic proxy
-    client = AsyncOpenAI(
-        base_url="https://llm-proxy.metaculus.com/proxy/openai/v1",
-        default_headers={
-            "Content-Type": "application/json",
-            "Authorization": f"Token {METACULUS_TOKEN}",
-        },
-        api_key="Fake API Key since openai requires this not to be NONE. This isn't used",
-        max_retries=2,
-    )
-
-    async with llm_rate_limiter:
-        collected_content = []
-        stream = await client.chat.completions.create(
-            model=model,
-            messages=[{"role": "user", "content": prompt}],
-            temperature=temperature,
-            stream=True,
+    async def _call_perplexity(self, question: str) -> str:
+        system_prompt = clean_indents(
+            """
+            You are an assistant to a superforecaster.
+            The superforecaster will give you a question they intend to forecast on.
+            To be a great assistant, you generate a concise but detailed rundown of the most relevant news, including if the question would resolve Yes or No based on current information.
+            You do not produce forecasts yourself.
+            """
         )
-
-        async for chunk in stream:
-            if chunk.choices and chunk.choices[0].delta.content is not None:
-                collected_content.append(chunk.choices[0].delta.content)
-
-    return "".join(collected_content)
-
-
-def run_research(question: str) -> str:
-    research = ""
-    if GET_NEWS == True:
-        if ASKNEWS_CLIENT_ID and ASKNEWS_SECRET:
-            research = call_asknews(question)
-        elif EXA_API_KEY:
-            research = call_exa_smart_searcher(question)
-        elif PERPLEXITY_API_KEY:
-            research = call_perplexity(question)
-        else:
-            raise ValueError("No API key provided")
-    else:
-        research = "No research done"
-
-    print(f"########################\nResearch Found:\n{research}\n########################")
-
-    return research
-
-def call_perplexity(question: str) -> str:
-    url = "https://api.perplexity.ai/chat/completions"
-    api_key = PERPLEXITY_API_KEY
-    headers = {
-        "accept": "application/json",
-        "authorization": f"Bearer {api_key}",
-        "content-type": "application/json",
-    }
-    payload = {
-        "model": "llama-3.1-sonar-huge-128k-online",
-        "messages": [
-            {
-                "role": "system",  # this is a system prompt designed to guide the perplexity assistant
-                "content": """
-                You are an assistant to a superforecaster.
-                The superforecaster will give you a question they intend to forecast on.
-                To be a great assistant, you generate a concise but detailed rundown of the most relevant news, including if the question would resolve Yes or No based on current information.
-                You do not produce forecasts yourself.
-                """,
-            },
-            {
-                "role": "user",  # this is the actual prompt we ask the perplexity assistant to answer
-                "content": question,
-            },
-        ],
-    }
-    response = requests.post(url=url, json=payload, headers=headers)
-    if not response.ok:
-        raise Exception(response.text)
-    content = response.json()["choices"][0]["message"]["content"]
-    return content
-
-def call_exa_smart_searcher(question: str) -> str:
-    if OPENAI_API_KEY is None:
-        searcher = forecasting_tools.ExaSearcher(
-            include_highlights=True,
-            num_results=10,
+        model = GeneralLlm(
+            model="perplexity/sonar-pro",  # perplexity/sonar is cheaper, but does only 1 search.
+            temperature=0.1,
+            system_prompt=system_prompt,
         )
-        highlights = asyncio.run(searcher.invoke_for_highlights_in_relevance_order(question))
-        prioritized_highlights = highlights[:10]
-        combined_highlights = ""
-        for i, highlight in enumerate(prioritized_highlights):
-            combined_highlights += f'[Highlight {i+1}]:\nTitle: {highlight.source.title}\nURL: {highlight.source.url}\nText: "{highlight.highlight_text}"\n\n'
-        response = combined_highlights
-    else:
-        searcher = forecasting_tools.SmartSearcher(
+        response = await model.invoke(question)
+        return response
+
+    async def _call_exa_smart_searcher(self, question: str) -> str:
+        """
+        SmartSearcher is a custom class that is a wrapper around a search on Exa.ai
+        """
+        searcher = SmartSearcher(
+            model=self._get_final_decision_llm(),
             temperature=0,
             num_searches_to_run=2,
             num_sites_per_search=10,
@@ -310,771 +103,253 @@ def call_exa_smart_searcher(question: str) -> str:
             "a concise but detailed rundown of the most relevant news, including if the question"
             "would resolve Yes or No based on current information. You do not produce forecasts yourself."
             f"\n\nThe question is: {question}"
-        )
-        response = asyncio.run(searcher.invoke(prompt))
-
-    return response
-
-def call_asknews(question: str) -> str:
-    """
-    Use the AskNews `news` endpoint to get news context for your query.
-    The full API reference can be found here: https://docs.asknews.app/en/reference#get-/v1/news/search
-    """
-    ask = AskNewsSDK(
-        client_id=ASKNEWS_CLIENT_ID, client_secret=ASKNEWS_SECRET, scopes=set(["news"])
-    )
-
-    # get the latest news related to the query (within the past 48 hours)
-    hot_response = ask.news.search_news(
-        query=question,  # your natural language query
-        n_articles=6,  # control the number of articles to include in the context, originally 5
-        return_type="both",
-        strategy="latest news",  # enforces looking at the latest news only
-    )
-
-    # get context from the "historical" database that contains a news archive going back to 2023
-    historical_response = ask.news.search_news(
-        query=question,
-        n_articles=10,
-        return_type="both",
-        strategy="news knowledge",  # looks for relevant news within the past 60 days
-    )
-
-    hot_articles = hot_response.as_dicts
-    historical_articles = historical_response.as_dicts
-    formatted_articles = "Here are the relevant news articles:\n\n"
-
-    if hot_articles:
-        hot_articles = [article.__dict__ for article in hot_articles]
-        hot_articles = sorted(hot_articles, key=lambda x: x["pub_date"], reverse=True)
-
-        for article in hot_articles:
-            pub_date = article["pub_date"].strftime("%B %d, %Y %I:%M %p")
-            formatted_articles += f"**{article['eng_title']}**\n{article['summary']}\nOriginal language: {article['language']}\nPublish date: {pub_date}\nSource:[{article['source_id']}]({article['article_url']})\n\n"
-
-    if historical_articles:
-        historical_articles = [article.__dict__ for article in historical_articles]
-        historical_articles = sorted(
-            historical_articles, key=lambda x: x["pub_date"], reverse=True
-        )
-
-        for article in historical_articles:
-            pub_date = article["pub_date"].strftime("%B %d, %Y %I:%M %p")
-            formatted_articles += f"**{article['eng_title']}**\n{article['summary']}\nOriginal language: {article['language']}\nPublish date: {pub_date}\nSource:[{article['source_id']}]({article['article_url']})\n\n"
-
-    if not hot_articles and not historical_articles:
-        formatted_articles += "No articles were found.\n\n"
-        return formatted_articles
-
-    return formatted_articles
-
-############### BINARY ###############
-# @title Binary prompt & functions
-
-# This section includes functionality for binary questions.
-
-BINARY_PROMPT_TEMPLATE = """
-You are a professional forecaster interviewing for a job.
-
-Your interview question is:
-{title}
-
-Question background:
-{background}
-
-
-This question's outcome will be determined by the specific criteria below. These criteria have not yet been satisfied:
-{resolution_criteria}
-
-{fine_print}
-
-
-Your research assistant says:
-{summary_report}
-
-Today is {today}.
-
-Before answering you write:
-(a) The time left until the outcome to the question is known.
-(b) The status quo outcome if nothing changed.
-(c) A brief description of a scenario that results in a No outcome.
-(d) A brief description of a scenario that results in a Yes outcome.
-
-You write your rationale remembering that good forecasters put extra weight on the status quo outcome since the world changes slowly most of the time.
-
-The last thing you write is your final answer as: "Probability: ZZ%", 0-100
-"""
-
-
-def extract_probability_from_response_as_percentage_not_decimal(
-    forecast_text: str,
-) -> float:
-    matches = re.findall(r"(\d+)%", forecast_text)
-    if matches:
-        # Return the last number found before a '%'
-        number = int(matches[-1])
-        number = min(99, max(1, number))  # clamp the number between 1 and 99
-        return number
-    else:
-        raise ValueError(f"Could not extract prediction from response: {forecast_text}")
-
-
-async def get_binary_gpt_prediction(
-    question_details: dict, num_runs: int
-) -> tuple[float, str]:
-
-    today = datetime.datetime.now().strftime("%Y-%m-%d")
-    title = question_details["title"]
-    resolution_criteria = question_details["resolution_criteria"]
-    background = question_details["description"]
-    fine_print = question_details["fine_print"]
-    question_type = question_details["type"]
-
-    summary_report = run_research(title)
-
-    content = BINARY_PROMPT_TEMPLATE.format(
-        title=title,
-        today=today,
-        background=background,
-        resolution_criteria=resolution_criteria,
-        fine_print=fine_print,
-        summary_report=summary_report,
-    )
-
-    async def get_rationale_and_probability(content: str) -> tuple[float, str]:
-        rationale = await call_llm(content)
-
-        probability = extract_probability_from_response_as_percentage_not_decimal(
-            rationale
-        )
-        comment = (
-            f"Extracted Probability: {probability}%\n\nGPT's Answer: "
-            f"{rationale}\n\n\n"
-        )
-        return probability, comment
-
-    probability_and_comment_pairs = await asyncio.gather(
-        *[get_rationale_and_probability(content) for _ in range(num_runs)]
-    )
-    comments = [pair[1] for pair in probability_and_comment_pairs]
-    final_comment_sections = [
-        f"## Rationale {i+1}\n{comment}" for i, comment in enumerate(comments)
-    ]
-    probabilities = [pair[0] for pair in probability_and_comment_pairs]
-    median_probability = float(np.median(probabilities)) / 100
-
-    final_comment = f"Median Probability: {median_probability}\n\n" + "\n\n".join(
-        final_comment_sections
-    )
-    return median_probability, final_comment
-
-
-####################### NUMERIC ###############
-# @title Numeric prompt & functions
-
-NUMERIC_PROMPT_TEMPLATE = """
-You are a professional forecaster interviewing for a job.
-
-Your interview question is:
-{title}
-
-Background:
-{background}
-
-{resolution_criteria}
-
-{fine_print}
-
-
-Your research assistant says:
-{summary_report}
-
-Today is {today}.
-
-{lower_bound_message}
-{upper_bound_message}
-
-
-Formatting Instructions:
-- Please notice the units requested (e.g. whether you represent a number as 1,000,000 or 1m).
-- Never use scientific notation.
-- Always start with a smaller number (more negative if negative) and then increase from there
-
-Before answering you write:
-(a) The time left until the outcome to the question is known.
-(b) The outcome if nothing changed.
-(c) The outcome if the current trend continued.
-(d) The expectations of experts and markets.
-(e) A brief description of an unexpected scenario that results in a low outcome.
-(f) A brief description of an unexpected scenario that results in a high outcome.
-
-You remind yourself that good forecasters are humble and set wide 90/10 confidence intervals to account for unknown unkowns.
-
-The last thing you write is your final answer as:
-"
-Percentile 10: XX
-Percentile 20: XX
-Percentile 40: XX
-Percentile 60: XX
-Percentile 80: XX
-Percentile 90: XX
-"
-"""
-
-
-def extract_percentiles_from_response(forecast_text: str) -> dict:
-
-    # Helper function that returns a list of tuples with numbers for all lines with Percentile
-    def extract_percentile_numbers(text) -> dict:
-        pattern = r"^.*(?:P|p)ercentile.*$"
-        number_pattern = r"-\s*(?:[^\d\-]*\s*)?(\d+(?:,\d{3})*(?:\.\d+)?)|(\d+(?:,\d{3})*(?:\.\d+)?)"
-        results = []
-
-        for line in text.split("\n"):
-            if re.match(pattern, line):
-                numbers = re.findall(number_pattern, line)
-                numbers_no_commas = [
-                    next(num for num in match if num).replace(",", "")
-                    for match in numbers
-                ]
-                numbers = [
-                    float(num) if "." in num else int(num)
-                    for num in numbers_no_commas
-                ]
-                if len(numbers) > 1:
-                    first_number = numbers[0]
-                    last_number = numbers[-1]
-                    # Check if the original line had a negative sign before the last number
-                    if "-" in line.split(":")[-1]:
-                        last_number = -abs(last_number)
-                    results.append((first_number, last_number))
-
-        # Convert results to dictionary
-        percentile_values = {}
-        for first_num, second_num in results:
-            key = first_num
-            percentile_values[key] = second_num
-
-        return percentile_values
-
-    percentile_values = extract_percentile_numbers(forecast_text)
-
-    if len(percentile_values) > 0:
-        return percentile_values
-    else:
-        raise ValueError(f"Could not extract prediction from response: {forecast_text}")
-
-
-def generate_continuous_cdf(
-    percentile_values: dict,
-    question_type: str,
-    open_upper_bound: bool,
-    open_lower_bound: bool,
-    upper_bound: float,
-    lower_bound: float,
-    zero_point: float | None,
-) -> list[float]:
-    """
-    Returns: list[float]: A list of 201 float values representing the CDF.
-    """
-
-    percentile_max = max(float(key) for key in percentile_values.keys())
-    percentile_min = min(float(key) for key in percentile_values.keys())
-    range_min = lower_bound
-    range_max = upper_bound
-    range_size = range_max - range_min
-    buffer = 1 if range_size > 100 else 0.01 * range_size
-
-    # Adjust any values that are exactly at the bounds
-    for percentile, value in list(percentile_values.items()):
-        if not open_lower_bound and value <= range_min + buffer:
-            percentile_values[percentile] = range_min + buffer
-        if not open_upper_bound and value >= range_max - buffer:
-            percentile_values[percentile] = range_max - buffer
-
-    # Set cdf values outside range
-    if open_upper_bound:
-        if range_max > percentile_values[percentile_max]:
-            percentile_values[int(100 - (0.5 * (100 - percentile_max)))] = range_max
-    else:
-        percentile_values[100] = range_max
-
-    # Set cdf values outside range
-    if open_lower_bound:
-        if range_min < percentile_values[percentile_min]:
-            percentile_values[int(0.5 * percentile_min)] = range_min
-    else:
-        percentile_values[0] = range_min
-
-    sorted_percentile_values = dict(sorted(percentile_values.items()))
-
-    # Normalize percentile keys
-    normalized_percentile_values = {}
-    for key, value in sorted_percentile_values.items():
-        percentile = float(key) / 100
-        normalized_percentile_values[percentile] = value
-
-
-    value_percentiles = {
-        value: key for key, value in normalized_percentile_values.items()
-    }
-
-    # function for log scaled questions
-    def generate_cdf_locations(range_min, range_max, zero_point):
-        if zero_point is None:
-            scale = lambda x: range_min + (range_max - range_min) * x
-        else:
-            deriv_ratio = (range_max - zero_point) / (range_min - zero_point)
-            scale = lambda x: range_min + (range_max - range_min) * (
-                deriv_ratio**x - 1
-            ) / (deriv_ratio - 1)
-        return [scale(x) for x in np.linspace(0, 1, 201)]
-
-    cdf_xaxis = generate_cdf_locations(range_min, range_max, zero_point)
-
-    def linear_interpolation(x_values, xy_pairs):
-        # Sort the xy_pairs by x-values
-        sorted_pairs = sorted(xy_pairs.items())
-
-        # Extract sorted x and y values
-        known_x = [pair[0] for pair in sorted_pairs]
-        known_y = [pair[1] for pair in sorted_pairs]
-
-        # Initialize the result list
-        y_values = []
-
-        for x in x_values:
-            # Check if x is exactly in the known x values
-            if x in known_x:
-                y_values.append(known_y[known_x.index(x)])
-            else:
-                # Find the indices of the two nearest known x-values
-                i = 0
-                while i < len(known_x) and known_x[i] < x:
-                    i += 1
-
-                list_index_2 = i
-
-                # If x is outside the range of known x-values, use the nearest endpoint
-                if i == 0:
-                    y_values.append(known_y[0])
-                elif i == len(known_x):
-                    y_values.append(known_y[-1])
-                else:
-                    # Perform linear interpolation
-                    x0, x1 = known_x[i - 1], known_x[i]
-                    y0, y1 = known_y[i - 1], known_y[i]
-
-                    # Linear interpolation formula
-                    y = y0 + (x - x0) * (y1 - y0) / (x1 - x0)
-                    y_values.append(y)
-
-        return y_values
-
-    continuous_cdf = linear_interpolation(cdf_xaxis, value_percentiles)
-    return continuous_cdf
-
-
-async def get_numeric_gpt_prediction(
-    question_details: dict, num_runs: int
-) -> tuple[list[float], str]:
-
-    today = datetime.datetime.now().strftime("%Y-%m-%d")
-    title = question_details["title"]
-    resolution_criteria = question_details["resolution_criteria"]
-    background = question_details["description"]
-    fine_print = question_details["fine_print"]
-    question_type = question_details["type"]
-    scaling = question_details["scaling"]
-    open_upper_bound = question_details["open_upper_bound"]
-    open_lower_bound = question_details["open_lower_bound"]
-    upper_bound = scaling["range_max"]
-    lower_bound = scaling["range_min"]
-    zero_point = scaling["zero_point"]
-
-    # Create messages about the bounds that are passed in the LLM prompt
-    if open_upper_bound:
-        upper_bound_message = ""
-    else:
-        upper_bound_message = f"The outcome can not be higher than {upper_bound}."
-    if open_lower_bound:
-        lower_bound_message = ""
-    else:
-        lower_bound_message = f"The outcome can not be lower than {lower_bound}."
-
-    summary_report = run_research(title)
-
-    content = NUMERIC_PROMPT_TEMPLATE.format(
-        title=title,
-        today=today,
-        background=background,
-        resolution_criteria=resolution_criteria,
-        fine_print=fine_print,
-        summary_report=summary_report,
-        lower_bound_message=lower_bound_message,
-        upper_bound_message=upper_bound_message,
-    )
-
-    async def ask_llm_to_get_cdf(content: str) -> tuple[list[float], str]:
-        rationale = await call_llm(content)
-        percentile_values = extract_percentiles_from_response(rationale)
-
-        comment = (
-            f"Extracted Percentile_values: {percentile_values}\n\nGPT's Answer: "
-            f"{rationale}\n\n\n"
-        )
-
-        cdf = generate_continuous_cdf(
-            percentile_values,
-            question_type,
-            open_upper_bound,
-            open_lower_bound,
-            upper_bound,
-            lower_bound,
-            zero_point,
-        )
-
-        return cdf, comment
-
-    cdf_and_comment_pairs = await asyncio.gather(
-        *[ask_llm_to_get_cdf(content) for _ in range(num_runs)]
-    )
-    comments = [pair[1] for pair in cdf_and_comment_pairs]
-    final_comment_sections = [
-        f"## Rationale {i+1}\n{comment}" for i, comment in enumerate(comments)
-    ]
-    cdfs: list[list[float]] = [pair[0] for pair in cdf_and_comment_pairs]
-    all_cdfs = np.array(cdfs)
-    median_cdf: list[float] = np.median(all_cdfs, axis=0).tolist()
-
-    final_comment = f"Median CDF: `{str(median_cdf)[:100]}...`\n\n" + "\n\n".join(
-        final_comment_sections
-    )
-    return median_cdf, final_comment
-
-
-########################## MULTIPLE CHOICE ###############
-# @title Multiple Choice prompt & functions
-
-MULTIPLE_CHOICE_PROMPT_TEMPLATE = """
-You are a professional forecaster interviewing for a job.
-
-Your interview question is:
-{title}
-
-The options are: {options}
-
-
-Background:
-{background}
-
-{resolution_criteria}
-
-{fine_print}
-
-
-Your research assistant says:
-{summary_report}
-
-Today is {today}.
-
-Before answering you write:
-(a) The time left until the outcome to the question is known.
-(b) The status quo outcome if nothing changed.
-(c) A description of an scenario that results in an unexpected outcome.
-
-You write your rationale remembering that (1) good forecasters put extra weight on the status quo outcome since the world changes slowly most of the time, and (2) good forecasters leave some moderate probability on most options to account for unexpected outcomes.
-
-The last thing you write is your final probabilities for the N options in this order {options} as:
-Option_A: Probability_A
-Option_B: Probability_B
-...
-Option_N: Probability_N
-"""
-
-
-def extract_option_probabilities_from_response(forecast_text: str, options) -> float:
-
-    # Helper function that returns a list of tuples with numbers for all lines with Percentile
-    def extract_option_probabilities(text):
-
-        # Number extraction pattern
-        number_pattern = r"-?\d+(?:,\d{3})*(?:\.\d+)?"
-
-        results = []
-
-        # Iterate through each line in the text
-        for line in text.split("\n"):
-            # Extract all numbers from the line
-            numbers = re.findall(number_pattern, line)
-            numbers_no_commas = [num.replace(",", "") for num in numbers]
-            # Convert strings to float or int
-            numbers = [
-                float(num) if "." in num else int(num) for num in numbers_no_commas
-            ]
-            # Add the tuple of numbers to results
-            if len(numbers) >= 1:
-                last_number = numbers[-1]
-                results.append(last_number)
-
-        return results
-
-    option_probabilities = extract_option_probabilities(forecast_text)
-
-    NUM_OPTIONS = len(options)
-
-    if len(option_probabilities) > 0:
-        # return the last NUM_OPTIONS items
-        return option_probabilities[-NUM_OPTIONS:]
-    else:
-        raise ValueError(f"Could not extract prediction from response: {forecast_text}")
-
-
-def generate_multiple_choice_forecast(options, option_probabilities) -> dict:
-    """
-    Returns: dict corresponding to the probabilities of each option.
-    """
-
-    # confirm that there is a probability for each option
-    if len(options) != len(option_probabilities):
-        raise ValueError(
-            f"Number of options ({len(options)}) does not match number of probabilities ({len(option_probabilities)})"
-        )
-
-    # Ensure we are using decimals
-    total_sum = sum(option_probabilities)
-    decimal_list = [x / total_sum for x in option_probabilities]
-
-    def normalize_list(float_list):
-        # Step 1: Clamp values
-        clamped_list = [max(min(x, 0.99), 0.01) for x in float_list]
-
-        # Step 2: Calculate the sum of all elements
-        total_sum = sum(clamped_list)
-
-        # Step 3: Normalize the list so that all elements add up to 1
-        normalized_list = [x / total_sum for x in clamped_list]
-
-        # Step 4: Adjust for any small floating-point errors
-        adjustment = 1.0 - sum(normalized_list)
-        normalized_list[-1] += adjustment
-
-        return normalized_list
-
-    normalized_option_probabilities = normalize_list(decimal_list)
-
-    probability_yes_per_category = {}
-    for i in range(len(options)):
-        probability_yes_per_category[options[i]] = normalized_option_probabilities[i]
-
-    return probability_yes_per_category
-
-
-async def get_multiple_choice_gpt_prediction(
-    question_details: dict,
-    num_runs: int,
-) -> tuple[dict[str, float], str]:
-
-    today = datetime.datetime.now().strftime("%Y-%m-%d")
-    title = question_details["title"]
-    resolution_criteria = question_details["resolution_criteria"]
-    background = question_details["description"]
-    fine_print = question_details["fine_print"]
-    question_type = question_details["type"]
-    options = question_details["options"]
-
-    summary_report = run_research(title)
-
-    content = MULTIPLE_CHOICE_PROMPT_TEMPLATE.format(
-        title=title,
-        today=today,
-        background=background,
-        resolution_criteria=resolution_criteria,
-        fine_print=fine_print,
-        summary_report=summary_report,
-        options=options,
-    )
-
-    async def ask_llm_for_multiple_choice_probabilities(
-        content: str,
-    ) -> tuple[dict[str, float], str]:
-        rationale = await call_llm(content)
-
-
-        option_probabilities = extract_option_probabilities_from_response(
-            rationale, options
-        )
-
-        comment = (
-            f"EXTRACTED_PROBABILITIES: {option_probabilities}\n\nGPT's Answer: "
-            f"{rationale}\n\n\n"
-        )
-
-        probability_yes_per_category = generate_multiple_choice_forecast(
-            options, option_probabilities
-        )
-        return probability_yes_per_category, comment
-
-    probability_yes_per_category_and_comment_pairs = await asyncio.gather(
-        *[ask_llm_for_multiple_choice_probabilities(content) for _ in range(num_runs)]
-    )
-    comments = [pair[1] for pair in probability_yes_per_category_and_comment_pairs]
-    final_comment_sections = [
-        f"## Rationale {i+1}\n{comment}" for i, comment in enumerate(comments)
-    ]
-    probability_yes_per_category_dicts: list[dict[str, float]] = [
-        pair[0] for pair in probability_yes_per_category_and_comment_pairs
-    ]
-    average_probability_yes_per_category: dict[str, float] = {}
-    for option in options:
-        probabilities_for_current_option: list[float] = [
-            dict[option] for dict in probability_yes_per_category_dicts
-        ]
-        average_probability_yes_per_category[option] = sum(
-            probabilities_for_current_option
-        ) / len(probabilities_for_current_option)
-
-    final_comment = (
-        f"Average Probability Yes Per Category: `{average_probability_yes_per_category}`\n\n"
-        + "\n\n".join(final_comment_sections)
-    )
-    return average_probability_yes_per_category, final_comment
-
-
-################### FORECASTING ###################
-def forecast_is_already_made(post_details: dict) -> bool:
-    """
-    Check if a forecast has already been made by looking at my_forecasts in the question data.
-
-    question.my_forecasts.latest.forecast_values has the following values for each question type:
-    Binary: [probability for no, probability for yes]
-    Numeric: [cdf value 1, cdf value 2, ..., cdf value 201]
-    Multiple Choice: [probability for option 1, probability for option 2, ...]
-    """
-    try:
-        forecast_values = post_details["question"]["my_forecasts"]["latest"][
-            "forecast_values"
-        ]
-        return forecast_values is not None
-    except Exception:
-        return False
-
-
-async def forecast_individual_question(
-    question_id: int,
-    post_id: int,
-    submit_prediction: bool,
-    num_runs_per_question: int,
-    skip_previously_forecasted_questions: bool,
-) -> str:
-    post_details = get_post_details(post_id)
-    question_details = post_details["question"]
-    title = question_details["title"]
-    question_type = question_details["type"]
-
-    summary_of_forecast = ""
-    summary_of_forecast += f"-----------------------------------------------\nQuestion: {title}\n"
-    summary_of_forecast += f"URL: https://www.metaculus.com/questions/{post_id}/\n"
-
-    if question_type == "multiple_choice":
-        options = question_details["options"]
-        summary_of_forecast += f"options: {options}\n"
-
-    if (
-        forecast_is_already_made(post_details)
-        and skip_previously_forecasted_questions == True
-    ):
-        summary_of_forecast += f"Skipped: Forecast already made\n"
-        return summary_of_forecast
-
-    if question_type == "binary":
-        forecast, comment = await get_binary_gpt_prediction(
-            question_details, num_runs_per_question
-        )
-    elif question_type == "numeric":
-        forecast, comment = await get_numeric_gpt_prediction(
-            question_details, num_runs_per_question
-        )
-    elif question_type == "multiple_choice":
-        forecast, comment = await get_multiple_choice_gpt_prediction(
-            question_details, num_runs_per_question
-        )
-    else:
-        raise ValueError(f"Unknown question type: {question_type}")
-
-    print(f"-----------------------------------------------\nPost {post_id} Question {question_id}:\n")
-    print(f"Forecast for post {post_id} (question {question_id}):\n{forecast}")
-    print(f"Comment for post {post_id} (question {question_id}):\n{comment}")
-
-    if question_type == "numeric":
-        summary_of_forecast += f"Forecast: {str(forecast)[:200]}...\n"
-    else:
-        summary_of_forecast += f"Forecast: {forecast}\n"
-
-    summary_of_forecast += f"Comment:\n```\n{comment[:200]}...\n```\n\n"
-
-    if submit_prediction == True:
-        forecast_payload = create_forecast_payload(forecast, question_type)
-        post_question_prediction(question_id, forecast_payload)
-        post_question_comment(post_id, comment)
-        summary_of_forecast += "Posted: Forecast was posted to Metaculus.\n"
-
-    return summary_of_forecast
-
-
-async def forecast_questions(
-    open_question_id_post_id: list[tuple[int, int]],
-    submit_prediction: bool,
-    num_runs_per_question: int,
-    skip_previously_forecasted_questions: bool,
-) -> None:
-    forecast_tasks = [
-        forecast_individual_question(
-            question_id,
-            post_id,
-            submit_prediction,
-            num_runs_per_question,
-            skip_previously_forecasted_questions,
-        )
-        for question_id, post_id in open_question_id_post_id
-    ]
-    forecast_summaries = await asyncio.gather(*forecast_tasks, return_exceptions=True)
-    print("\n", "#" * 100, "\nForecast Summaries\n", "#" * 100)
-
-    errors = []
-    for question_id_post_id, forecast_summary in zip(
-        open_question_id_post_id, forecast_summaries
-    ):
-        question_id, post_id = question_id_post_id
-        if isinstance(forecast_summary, Exception):
-            print(
-                f"-----------------------------------------------\nPost {post_id} Question {question_id}:\nError: {forecast_summary.__class__.__name__} {forecast_summary}\nURL: https://www.metaculus.com/questions/{post_id}/\n"
+        )  # You can ask the searcher to filter by date, exclude/include a domain, and run specific searches for finding sources vs finding highlights within a source
+        response = await searcher.invoke(prompt)
+        return response
+
+    def _get_final_decision_llm(self) -> GeneralLlm:
+        model = None
+        if os.getenv("OPENAI_API_KEY"):
+            model = GeneralLlm(model="gpt-4o", temperature=0.3)
+        elif os.getenv("ANTHROPIC_API_KEY"):
+            model = GeneralLlm(
+                model="claude-3-5-sonnet-20241022", temperature=0.3
             )
-            errors.append(forecast_summary)
+        elif os.getenv("OPENROUTER_API_KEY"):
+            model = GeneralLlm(
+                model="openrouter/openai/gpt-4o", temperature=0.3
+            )
+        elif os.getenv("METACULUS_TOKEN"):
+            model = GeneralLlm(model="metaculus/gpt-4o", temperature=0.3)
         else:
-            print(forecast_summary)
+            raise ValueError("No API key for final_decision_llm found")
+        return model
 
-    if errors:
-        print("-----------------------------------------------\nErrors:\n")
-        error_message = f"Errors were encountered: {errors}"
-        print(error_message)
-        raise RuntimeError(error_message)
+    async def _run_forecast_on_binary(
+        self, question: BinaryQuestion, research: str
+    ) -> ReasonedPrediction[float]:
+        prompt = clean_indents(
+            f"""
+            You are a professional forecaster interviewing for a job.
+
+            Your interview question is:
+            {question.question_text}
+
+            Question background:
+            {question.background_info}
 
 
+            This question's outcome will be determined by the specific criteria below. These criteria have not yet been satisfied:
+            {question.resolution_criteria}
+
+            {question.fine_print}
 
 
-######################## FINAL RUN #########################
-if __name__ == "__main__":
-    if USE_EXAMPLE_QUESTIONS:
-        open_question_id_post_id = EXAMPLE_QUESTIONS
-    else:
-        open_question_id_post_id = get_open_question_ids_from_tournament()
+            Your research assistant says:
+            {research}
 
-    asyncio.run(
-        forecast_questions(
-            open_question_id_post_id,
-            SUBMIT_PREDICTION,
-            NUM_RUNS_PER_QUESTION,
-            SKIP_PREVIOUSLY_FORECASTED_QUESTIONS,
+            Today is {datetime.now().strftime("%Y-%m-%d")}.
+
+            Before answering you write:
+            (a) The time left until the outcome to the question is known.
+            (b) The status quo outcome if nothing changed.
+            (c) A brief description of a scenario that results in a No outcome.
+            (d) A brief description of a scenario that results in a Yes outcome.
+
+            You write your rationale remembering that good forecasters put extra weight on the status quo outcome since the world changes slowly most of the time.
+
+            The last thing you write is your final answer as: "Probability: ZZ%", 0-100
+            """
         )
+        reasoning = await self._get_final_decision_llm().invoke(prompt)
+        prediction: float = PredictionExtractor.extract_last_percentage_value(
+            reasoning, max_prediction=1, min_prediction=0
+        )
+        logger.info(
+            f"Forecasted {question.page_url} as {prediction} with reasoning:\n{reasoning}"
+        )
+        return ReasonedPrediction(
+            prediction_value=prediction, reasoning=reasoning
+        )
+
+    async def _run_forecast_on_multiple_choice(
+        self, question: MultipleChoiceQuestion, research: str
+    ) -> ReasonedPrediction[PredictedOptionList]:
+        prompt = clean_indents(
+            f"""
+            You are a professional forecaster interviewing for a job.
+
+            Your interview question is:
+            {question.question_text}
+
+            The options are: {question.options}
+
+
+            Background:
+            {question.background_info}
+
+            {question.resolution_criteria}
+
+            {question.fine_print}
+
+
+            Your research assistant says:
+            {research}
+
+            Today is {datetime.now().strftime("%Y-%m-%d")}.
+
+            Before answering you write:
+            (a) The time left until the outcome to the question is known.
+            (b) The status quo outcome if nothing changed.
+            (c) A description of an scenario that results in an unexpected outcome.
+
+            You write your rationale remembering that (1) good forecasters put extra weight on the status quo outcome since the world changes slowly most of the time, and (2) good forecasters leave some moderate probability on most options to account for unexpected outcomes.
+
+            The last thing you write is your final probabilities for the N options in this order {question.options} as:
+            Option_A: Probability_A
+            Option_B: Probability_B
+            ...
+            Option_N: Probability_N
+            """
+        )
+        reasoning = await self._get_final_decision_llm().invoke(prompt)
+        prediction: PredictedOptionList = (
+            PredictionExtractor.extract_option_list_with_percentage_afterwards(
+                reasoning, question.options
+            )
+        )
+        logger.info(
+            f"Forecasted {question.page_url} as {prediction} with reasoning:\n{reasoning}"
+        )
+        return ReasonedPrediction(
+            prediction_value=prediction, reasoning=reasoning
+        )
+
+    async def _run_forecast_on_numeric(
+        self, question: NumericQuestion, research: str
+    ) -> ReasonedPrediction[NumericDistribution]:
+        upper_bound_message, lower_bound_message = (
+            self._create_upper_and_lower_bound_messages(question)
+        )
+        prompt = clean_indents(
+            f"""
+            You are a professional forecaster interviewing for a job.
+
+            Your interview question is:
+            {question.question_text}
+
+            Background:
+            {question.background_info}
+
+            {question.resolution_criteria}
+
+            {question.fine_print}
+
+
+            Your research assistant says:
+            {research}
+
+            Today is {datetime.now().strftime("%Y-%m-%d")}.
+
+            {lower_bound_message}
+            {upper_bound_message}
+
+            Formatting Instructions:
+            - Please notice the units requested (e.g. whether you represent a number as 1,000,000 or 1m).
+            - Never use scientific notation.
+            - Always start with a smaller number (more negative if negative) and then increase from there
+
+            Before answering you write:
+            (a) The time left until the outcome to the question is known.
+            (b) The outcome if nothing changed.
+            (c) The outcome if the current trend continued.
+            (d) The expectations of experts and markets.
+            (e) A brief description of an unexpected scenario that results in a low outcome.
+            (f) A brief description of an unexpected scenario that results in a high outcome.
+
+            You remind yourself that good forecasters are humble and set wide 90/10 confidence intervals to account for unknown unknowns.
+
+            The last thing you write is your final answer as:
+            "
+            Percentile 10: XX
+            Percentile 20: XX
+            Percentile 40: XX
+            Percentile 60: XX
+            Percentile 80: XX
+            Percentile 90: XX
+            "
+            """
+        )
+        reasoning = await self._get_final_decision_llm().invoke(prompt)
+        prediction: NumericDistribution = (
+            PredictionExtractor.extract_numeric_distribution_from_list_of_percentile_number_and_probability(
+                reasoning, question
+            )
+        )
+        logger.info(
+            f"Forecasted {question.page_url} as {prediction.declared_percentiles} with reasoning:\n{reasoning}"
+        )
+        return ReasonedPrediction(
+            prediction_value=prediction, reasoning=reasoning
+        )
+
+    def _create_upper_and_lower_bound_messages(
+        self, question: NumericQuestion
+    ) -> tuple[str, str]:
+        if question.open_upper_bound:
+            upper_bound_message = ""
+        else:
+            upper_bound_message = (
+                f"The outcome can not be higher than {question.upper_bound}."
+            )
+        if question.open_lower_bound:
+            lower_bound_message = ""
+        else:
+            lower_bound_message = (
+                f"The outcome can not be lower than {question.lower_bound}."
+            )
+        return upper_bound_message, lower_bound_message
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description="Run the Q1TemplateBot forecasting system")
+    parser.add_argument(
+        "--mode",
+        type=str,
+        choices=["tournament", "quarterly_cup", "test_questions"],
+        default="tournament",
+        help="Specify the run mode (default: tournament)"
     )
+    args = parser.parse_args()
+    run_mode: Literal["tournament", "quarterly_cup", "test_questions"] = args.mode
+    assert run_mode in ["tournament", "quarterly_cup", "test_questions"], "Invalid run mode"
+
+    template_bot = Q1TemplateBot(
+        research_reports_per_question=1,
+        predictions_per_research_report=5,
+        use_research_summary_to_forecast=False,
+        publish_reports_to_metaculus=True,
+        folder_to_save_reports_to=None,
+        skip_previously_forecasted_questions=True,
+    )
+
+    if run_mode == "tournament":
+        Q4_2024_AI_BENCHMARKING_ID = 32506
+        Q1_2025_AI_BENCHMARKING_ID = 32627
+        asyncio.run(template_bot.forecast_on_tournament(Q1_2025_AI_BENCHMARKING_ID))
+    elif run_mode == "quarterly_cup":
+        # The quarterly cup is a good way to test the bot's performance on regularly open questions. You can also use AXC_2025_TOURNAMENT_ID = 32564
+        Q1_2025_QUARTERLY_CUP_ID = 32630
+        template_bot.skip_previously_forecasted_questions = False
+        asyncio.run(template_bot.forecast_on_tournament(Q1_2025_QUARTERLY_CUP_ID))
+    elif run_mode == "test_questions":
+        # Example questions are a good way to test the bot's performance on a single question
+        EXAMPLE_QUESTIONS = [
+            "https://www.metaculus.com/questions/578/human-extinction-by-2100/",  # Human Extinction - Binary
+            "https://www.metaculus.com/questions/14333/age-of-oldest-human-as-of-2100/",  # Age of Oldest Human - Numeric
+            "https://www.metaculus.com/questions/22427/number-of-new-leading-ai-labs/",  # Number of New Leading AI Labs - Multiple Choice
+        ]
+        template_bot.skip_previously_forecasted_questions = False
+        questions = [MetaculusApi.get_question_by_url(question_url) for question_url in EXAMPLE_QUESTIONS]
+        for question in questions:
+            asyncio.run(template_bot.forecast_questions(questions))
