+++ pyproject.toml
@@ -14,7 +14,7 @@ asknews = "^0.9.1"
 numpy = "^1.26.0"
 openai = "^1.57.4"
 python-dotenv = "^1.0.1"
-forecasting-tools = "^0.2.10"
+forecasting-tools = "^0.2.13"
 matplotlib = "^3.10.0"
